M4 L5 UPDATE

+++ module4/lesson05_orm/exercise_product_manager/src/main/java/com/product/model/Product.java
@@ -8,16 +8,16 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Long id;
     private String name;
-    private double price;
+    private Double price;
     private String manufacture;
     private String description;
 
     public Product() {
     }
 
-    public Product(int id, String name, double price, String manufacture, String description) {
+    public Product(Long id, String name, double price, String manufacture, String description) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -25,11 +25,11 @@ public class Product {
         this.description = description;
     }
 
-    public long getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -41,11 +41,11 @@ public class Product {
         this.name = name;
     }
 
-    public double getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
